refactor(middleware): split session.rs by responsibility (#4312)
Decompose the 1763-line session.rs into responsibility-based submodules
under session/ following Rust 2024 Edition module conventions. Public
API preserved via pub use re-exports.

Fixes #4312
