+++ crates/reinhardt-middleware/src/session.rs
@@ -2,733 +2,48 @@
 //!
 //! Provides enhanced session management functionality.
 //! Supports various backends including Cookie, Redis, and database.
-
-use async_trait::async_trait;
-#[allow(deprecated)]
-use reinhardt_conf::Settings;
-use reinhardt_di::{DiError, DiResult, Injectable, InjectionContext};
-use reinhardt_http::{Handler, Middleware, Request, Response, Result};
-use serde::{Deserialize, Serialize};
-use std::collections::HashMap;
-use std::sync::{Arc, RwLock};
-use std::time::{Duration, SystemTime};
-use uuid::Uuid;
-
-/// Newtype wrapper for session ID stored in request extensions.
-///
-/// Handlers can retrieve the current session ID from the request
-/// extensions without parsing cookies manually.
-///
-/// # Example
-///
-/// ```rust,ignore
-/// fn handle(&self, request: Request) -> Result<Response> {
-///     if let Some(session_id) = request.extensions.get::<SessionId>() {
-///         println!("Session: {}", session_id.as_str());
-///     }
-///     // ...
-/// }
-/// ```
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct SessionId(String);
-
-impl SessionId {
-	/// Create a new `SessionId` from the given string.
-	pub fn new(id: String) -> Self {
-		Self(id)
-	}
-
-	/// Returns the session ID as a string slice.
-	pub fn as_str(&self) -> &str {
-		&self.0
-	}
-}
-
-impl AsRef<str> for SessionId {
-	fn as_ref(&self) -> &str {
-		self.as_str()
-	}
-}
-
-impl std::fmt::Display for SessionId {
-	fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-		f.write_str(self.as_str())
-	}
-}
-
-/// Shared, mutable handle to the session ID that the middleware will write
-/// to the response `Set-Cookie` header.
-///
-/// Stored in request extensions by `SessionMiddleware`. Handlers that rotate
-/// the session ID (e.g., for session-fixation prevention on login) MUST
-/// either call `SessionData::regenerate_id` (which updates this holder
-/// transparently) or write to it directly via `set`. Otherwise the cookie
-/// returned to the client points at a session ID that no longer exists in
-/// the store. See #3827.
-#[derive(Debug, Clone)]
-pub struct ActiveSessionId(Arc<RwLock<String>>);
-
-impl ActiveSessionId {
-	/// Create an `ActiveSessionId` initialised to `id`.
-	pub fn new(id: String) -> Self {
-		Self(Arc::new(RwLock::new(id)))
-	}
-
-	/// Read the current session ID.
-	pub fn get(&self) -> String {
-		self.0.read().unwrap_or_else(|e| e.into_inner()).clone()
-	}
-
-	/// Replace the session ID. Call after rotating the underlying
-	/// `SessionData::id` so the middleware's `Set-Cookie` matches the
-	/// store entry.
-	pub fn set(&self, id: String) {
-		*self.0.write().unwrap_or_else(|e| e.into_inner()) = id;
-	}
-}
-
-/// Newtype wrapper for the configured session cookie name.
-///
-/// Stored in request extensions by `SessionMiddleware` so that
-/// `Injectable` implementations can retrieve the configured cookie name
-/// instead of hardcoding it.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct SessionCookieName(String);
-
-impl SessionCookieName {
-	/// Create a new `SessionCookieName`.
-	pub fn new(name: String) -> Self {
-		Self(name)
-	}
-
-	/// Returns the cookie name as a string slice.
-	pub fn as_str(&self) -> &str {
-		&self.0
-	}
-}
-
-/// Session data
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[non_exhaustive]
-pub struct SessionData {
-	/// Session ID
-	pub id: String,
-	/// Data
-	pub data: HashMap<String, serde_json::Value>,
-	/// Creation timestamp
-	pub created_at: SystemTime,
-	/// Last access timestamp
-	pub last_accessed: SystemTime,
-	/// Expiration timestamp
-	pub expires_at: SystemTime,
-	/// Back-reference to the request-scoped active session ID holder.
-	///
-	/// Populated by `SessionData::inject` from the request extensions; used by
-	/// `regenerate_id` to keep the middleware's `Set-Cookie` value in sync
-	/// with the rotated session ID. Never serialized — sessions persisted to a
-	/// store carry only the data they own. See #3827.
-	///
-	/// Defaults to `None`; callers constructing `SessionData` literally outside
-	/// the middleware (tests, fixtures) can leave it `None` because rotation
-	/// only matters when the session is actively wired into a live request.
-	#[serde(skip)]
-	pub id_holder: Option<ActiveSessionId>,
-}
-
-impl SessionData {
-	/// Create a new session
-	pub fn new(ttl: Duration) -> Self {
-		let now = SystemTime::now();
-		Self {
-			id: Uuid::new_v4().to_string(),
-			data: HashMap::new(),
-			created_at: now,
-			last_accessed: now,
-			expires_at: now + ttl,
-			id_holder: None,
-		}
-	}
-
-	/// Rotate the session ID (e.g., after authentication, to prevent session
-	/// fixation). Updates both `self.id` and the request-scoped
-	/// [`ActiveSessionId`] so that `SessionMiddleware` writes the new ID to
-	/// the response cookie.
-	///
-	/// Returns the previous ID so callers can delete the stale entry from
-	/// the store.
-	///
-	/// See #3827.
-	pub fn regenerate_id(&mut self) -> String {
-		let old_id = std::mem::replace(&mut self.id, Uuid::now_v7().to_string());
-		if let Some(holder) = &self.id_holder {
-			holder.set(self.id.clone());
-		}
-		old_id
-	}
-
-	/// Check if session is valid
-	fn is_valid(&self) -> bool {
-		SystemTime::now() < self.expires_at
-	}
-
-	/// Update last access timestamp
-	pub fn touch(&mut self, ttl: Duration) {
-		let now = SystemTime::now();
-		self.last_accessed = now;
-		self.expires_at = now + ttl;
-	}
-
-	/// Get a value
-	pub fn get<T>(&self, key: &str) -> Option<T>
-	where
-		T: for<'de> Deserialize<'de>,
-	{
-		self.data
-			.get(key)
-			.and_then(|v| serde_json::from_value(v.clone()).ok())
-	}
-
-	/// Set a value
-	pub fn set<T>(&mut self, key: String, value: T) -> Result<()>
-	where
-		T: Serialize,
-	{
-		self.data.insert(
-			key,
-			serde_json::to_value(value)
-				.map_err(|e| reinhardt_core::exception::Error::Serialization(e.to_string()))?,
-		);
-		Ok(())
-	}
-
-	/// Delete a value
-	pub fn delete(&mut self, key: &str) {
-		self.data.remove(key);
-	}
-
-	/// Check if a key exists
-	pub fn contains_key(&self, key: &str) -> bool {
-		self.data.contains_key(key)
-	}
-
-	/// Clear the session
-	pub fn clear(&mut self) {
-		self.data.clear();
-	}
-}
-
-/// Session store with automatic lazy eviction of expired sessions
-///
-/// Performs periodic cleanup of expired sessions to prevent unbounded
-/// memory growth. Cleanup runs automatically when the session count
-/// exceeds a configurable threshold.
-#[derive(Debug, Default)]
-pub struct SessionStore {
-	/// Sessions
-	sessions: RwLock<HashMap<String, SessionData>>,
-	/// Maximum number of sessions before triggering automatic cleanup
-	max_sessions_before_cleanup: std::sync::atomic::AtomicUsize,
-}
-
-impl SessionStore {
-	/// Default cleanup threshold: trigger cleanup when session count exceeds 10,000
-	const DEFAULT_CLEANUP_THRESHOLD: usize = 10_000;
-
-	/// Create a new store
-	pub fn new() -> Self {
-		Self {
-			sessions: RwLock::new(HashMap::new()),
-			max_sessions_before_cleanup: std::sync::atomic::AtomicUsize::new(
-				Self::DEFAULT_CLEANUP_THRESHOLD,
-			),
-		}
-	}
-
-	/// Get a session
-	pub fn get(&self, id: &str) -> Option<SessionData> {
-		let sessions = self.sessions.read().unwrap_or_else(|e| e.into_inner());
-		sessions.get(id).cloned()
-	}
-
-	/// Save a session, with automatic cleanup when threshold is exceeded
-	pub fn save(&self, session: SessionData) {
-		let mut sessions = self.sessions.write().unwrap_or_else(|e| e.into_inner());
-		sessions.insert(session.id.clone(), session);
-
-		// Lazy eviction: clean up expired sessions when threshold is exceeded
-		let threshold = self
-			.max_sessions_before_cleanup
-			.load(std::sync::atomic::Ordering::Relaxed);
-		if sessions.len() > threshold {
-			sessions.retain(|_, s| s.is_valid());
-		}
-	}
-
-	/// Delete a session
-	pub fn delete(&self, id: &str) {
-		let mut sessions = self.sessions.write().unwrap_or_else(|e| e.into_inner());
-		sessions.remove(id);
-	}
-
-	/// Clean up expired sessions
-	pub fn cleanup(&self) {
-		let mut sessions = self.sessions.write().unwrap_or_else(|e| e.into_inner());
-		sessions.retain(|_, session| session.is_valid());
-	}
-
-	/// Clear the store
-	pub fn clear(&self) {
-		let mut sessions = self.sessions.write().unwrap_or_else(|e| e.into_inner());
-		sessions.clear();
-	}
-
-	/// Get the number of sessions
-	pub fn len(&self) -> usize {
-		let sessions = self.sessions.read().unwrap_or_else(|e| e.into_inner());
-		sessions.len()
-	}
-
-	/// Check if the store is empty
-	pub fn is_empty(&self) -> bool {
-		let sessions = self.sessions.read().unwrap_or_else(|e| e.into_inner());
-		sessions.is_empty()
-	}
-}
-
-/// Async trait for pluggable session storage backends.
-///
-/// Implement this trait to integrate any async-capable session store
-/// (e.g. Redis, DynamoDB, PostgreSQL) with the session middleware layer.
-///
-/// # Example
-///
-/// ```rust,ignore
-/// use std::time::Duration;
-/// use reinhardt_middleware::session::{AsyncSessionBackend, SessionData};
-/// use reinhardt_http::Result;
-///
-/// struct MyBackend;
-///
-/// #[async_trait::async_trait]
-/// impl AsyncSessionBackend for MyBackend {
-///     async fn load(&self, id: &str) -> Result<Option<SessionData>> { Ok(None) }
-///     async fn save(&self, session: &SessionData) -> Result<()> { Ok(()) }
-///     async fn destroy(&self, id: &str) -> Result<()> { Ok(()) }
-///     async fn touch(&self, id: &str, ttl: Duration) -> Result<()> { Ok(()) }
-/// }
-/// ```
-#[async_trait]
-pub trait AsyncSessionBackend: Send + Sync {
-	/// Load a session by ID. Returns `None` if the session does not exist
-	/// or has expired.
-	async fn load(&self, id: &str) -> Result<Option<SessionData>>;
-
-	/// Persist a session (insert or update).
-	async fn save(&self, session: &SessionData) -> Result<()>;
-
-	/// Remove a session by ID.
-	async fn destroy(&self, id: &str) -> Result<()>;
-
-	/// Refresh the TTL of an existing session without rewriting the full payload.
-	async fn touch(&self, id: &str, ttl: Duration) -> Result<()>;
-}
-
-/// Session configuration
-#[non_exhaustive]
-#[derive(Debug, Clone)]
-pub struct SessionConfig {
-	/// Cookie name
-	pub cookie_name: String,
-	/// Session TTL
-	pub ttl: Duration,
-	/// HTTPS-only cookie
-	pub secure: bool,
-	/// HttpOnly flag
-	pub http_only: bool,
-	/// SameSite attribute
-	pub same_site: Option<String>,
-	/// Domain
-	pub domain: Option<String>,
-	/// Path
-	pub path: String,
-}
-
-impl SessionConfig {
-	/// Create a new configuration
-	///
-	/// # Examples
-	///
-	/// ```
-	/// use std::time::Duration;
-	/// use reinhardt_middleware::session::SessionConfig;
-	///
-	/// let config = SessionConfig::new("sessionid".to_string(), Duration::from_secs(3600));
-	/// assert_eq!(config.cookie_name, "sessionid");
-	/// assert_eq!(config.ttl, Duration::from_secs(3600));
-	/// ```
-	pub fn new(cookie_name: String, ttl: Duration) -> Self {
-		Self {
-			cookie_name,
-			ttl,
-			secure: true,
-			http_only: true,
-			same_site: Some("Lax".to_string()),
-			domain: None,
-			path: "/".to_string(),
-		}
-	}
-
-	/// Enable secure cookie
-	///
-	/// # Examples
-	///
-	/// ```
-	/// use std::time::Duration;
-	/// use reinhardt_middleware::session::SessionConfig;
-	///
-	/// let config = SessionConfig::new("sessionid".to_string(), Duration::from_secs(3600))
-	///     .with_secure();
-	/// assert!(config.secure);
-	/// ```
-	pub fn with_secure(mut self) -> Self {
-		self.secure = true;
-		self
-	}
-
-	/// Set HttpOnly flag
-	///
-	/// # Examples
-	///
-	/// ```
-	/// use std::time::Duration;
-	/// use reinhardt_middleware::session::SessionConfig;
-	///
-	/// let config = SessionConfig::new("sessionid".to_string(), Duration::from_secs(3600))
-	///     .with_http_only(false);
-	/// assert!(!config.http_only);
-	/// ```
-	pub fn with_http_only(mut self, http_only: bool) -> Self {
-		self.http_only = http_only;
-		self
-	}
-
-	/// Set SameSite attribute
-	///
-	/// # Examples
-	///
-	/// ```
-	/// use std::time::Duration;
-	/// use reinhardt_middleware::session::SessionConfig;
-	///
-	/// let config = SessionConfig::new("sessionid".to_string(), Duration::from_secs(3600))
-	///     .with_same_site("Strict".to_string());
-	/// ```
-	pub fn with_same_site(mut self, same_site: String) -> Self {
-		self.same_site = Some(same_site);
-		self
-	}
-
-	/// Set domain
-	///
-	/// # Examples
-	///
-	/// ```
-	/// use std::time::Duration;
-	/// use reinhardt_middleware::session::SessionConfig;
-	///
-	/// let config = SessionConfig::new("sessionid".to_string(), Duration::from_secs(3600))
-	///     .with_domain("example.com".to_string());
-	/// ```
-	pub fn with_domain(mut self, domain: String) -> Self {
-		self.domain = Some(domain);
-		self
-	}
-
-	/// Set path
-	///
-	/// # Examples
-	///
-	/// ```
-	/// use std::time::Duration;
-	/// use reinhardt_middleware::session::SessionConfig;
-	///
-	/// let config = SessionConfig::new("sessionid".to_string(), Duration::from_secs(3600))
-	///     .with_path("/app".to_string());
-	/// assert_eq!(config.path, "/app");
-	/// ```
-	pub fn with_path(mut self, path: String) -> Self {
-		self.path = path;
-		self
-	}
-
-	/// Create a `SessionConfig` from application `Settings`
-	///
-	/// Maps `Settings.core.security.session_cookie_secure` to `SessionConfig.secure`.
-	///
-	/// # Examples
-	///
-	/// ```
-	/// use reinhardt_conf::Settings;
-	/// use reinhardt_middleware::session::SessionConfig;
-	///
-	/// #[allow(deprecated)]
-	/// let settings = Settings::default();
-	/// #[allow(deprecated)]
-	/// let config = SessionConfig::from_settings(&settings);
-	/// assert!(!config.secure);
-	/// ```
-	#[allow(deprecated)] // Settings is deprecated in favor of composable fragments
-	pub fn from_settings(settings: &Settings) -> Self {
-		Self {
-			secure: settings.core.security.session_cookie_secure,
-			..Self::default()
-		}
-	}
-}
-
-impl Default for SessionConfig {
-	fn default() -> Self {
-		Self::new("sessionid".to_string(), Duration::from_secs(3600))
-	}
-}
-
-/// Session middleware
-///
-/// # Examples
-///
-/// ```
-/// use std::sync::Arc;
-/// use std::time::Duration;
-/// use reinhardt_middleware::session::{SessionMiddleware, SessionConfig};
-/// use reinhardt_http::{Handler, Middleware, Request, Response};
-/// use hyper::{StatusCode, Method, Version, HeaderMap};
-/// use bytes::Bytes;
-///
-/// struct TestHandler;
-///
-/// #[async_trait::async_trait]
-/// impl Handler for TestHandler {
-///     async fn handle(&self, _request: Request) -> reinhardt_core::exception::Result<Response> {
-///         Ok(Response::new(StatusCode::OK).with_body(Bytes::from("OK")))
-///     }
-/// }
-///
-/// # tokio_test::block_on(async {
-/// let config = SessionConfig::new("sessionid".to_string(), Duration::from_secs(3600));
-/// let middleware = SessionMiddleware::new(config);
-/// let handler = Arc::new(TestHandler);
-///
-/// let request = Request::builder()
-///     .method(Method::GET)
-///     .uri("/api/data")
-///     .version(Version::HTTP_11)
-///     .headers(HeaderMap::new())
-///     .body(Bytes::new())
-///     .build()
-///     .unwrap();
-///
-/// let response = middleware.process(request, handler).await.unwrap();
-/// assert_eq!(response.status, StatusCode::OK);
-/// # });
-/// ```
-pub struct SessionMiddleware {
-	config: SessionConfig,
-	store: Arc<SessionStore>,
-}
-
-impl SessionMiddleware {
-	/// Create a new session middleware
-	///
-	/// # Examples
-	///
-	/// ```
-	/// use std::time::Duration;
-	/// use reinhardt_middleware::session::{SessionMiddleware, SessionConfig};
-	///
-	/// let config = SessionConfig::new("sessionid".to_string(), Duration::from_secs(3600));
-	/// let middleware = SessionMiddleware::new(config);
-	/// ```
-	pub fn new(config: SessionConfig) -> Self {
-		Self {
-			config,
-			store: Arc::new(SessionStore::new()),
-		}
-	}
-
-	/// Create a `SessionMiddleware` from application `Settings`
-	///
-	/// # Examples
-	///
-	/// ```
-	/// use reinhardt_conf::Settings;
-	/// use reinhardt_middleware::session::SessionMiddleware;
-	///
-	/// #[allow(deprecated)]
-	/// let settings = Settings::default();
-	/// #[allow(deprecated)]
-	/// let middleware = SessionMiddleware::from_settings(&settings);
-	/// ```
-	#[allow(deprecated)] // Settings is deprecated in favor of composable fragments
-	pub fn from_settings(settings: &Settings) -> Self {
-		Self::new(SessionConfig::from_settings(settings))
-	}
-
-	/// Create with default configuration
-	pub fn with_defaults() -> Self {
-		Self::new(SessionConfig::default())
-	}
-
-	/// Create from an existing Arc-wrapped session store
-	///
-	/// This is provided for cases where you already have an `Arc<SessionStore>`.
-	/// In most cases, you should use `new()` instead, which creates the store internally.
-	pub fn from_arc(config: SessionConfig, store: Arc<SessionStore>) -> Self {
-		Self { config, store }
-	}
-
-	/// Get a reference to the session store
-	///
-	/// # Examples
-	///
-	/// ```
-	/// use std::time::Duration;
-	/// use reinhardt_middleware::session::{SessionMiddleware, SessionConfig};
-	///
-	/// let middleware = SessionMiddleware::new(
-	///     SessionConfig::new("sessionid".to_string(), Duration::from_secs(3600))
-	/// );
-	///
-	/// // Access the store
-	/// let store = middleware.store();
-	/// assert_eq!(store.len(), 0);
-	/// ```
-	pub fn store(&self) -> &SessionStore {
-		&self.store
-	}
-
-	/// Get a cloned Arc of the store (for cases where you need ownership)
-	///
-	/// In most cases, you should use `store()` instead to get a reference.
-	pub fn store_arc(&self) -> Arc<SessionStore> {
-		Arc::clone(&self.store)
-	}
-
-	/// Get session ID from request
-	fn get_session_id(&self, request: &Request) -> Option<String> {
-		if let Some(cookie_header) = request.headers.get(hyper::header::COOKIE)
-			&& let Ok(cookie_str) = cookie_header.to_str()
-		{
-			for cookie in cookie_str.split(';') {
-				let parts: Vec<&str> = cookie.trim().splitn(2, '=').collect();
-				if parts.len() == 2 && parts[0] == self.config.cookie_name {
-					return Some(parts[1].to_string());
-				}
-			}
-		}
-		None
-	}
-
-	/// Build Set-Cookie header
-	fn build_cookie_header(&self, session_id: &str) -> String {
-		let mut parts = vec![format!("{}={}", self.config.cookie_name, session_id)];
-
-		parts.push(format!("Path={}", self.config.path));
-
-		if let Some(domain) = &self.config.domain {
-			parts.push(format!("Domain={}", domain));
-		}
-
-		if self.config.http_only {
-			parts.push("HttpOnly".to_string());
-		}
-
-		if self.config.secure {
-			parts.push("Secure".to_string());
-		}
-
-		if let Some(same_site) = &self.config.same_site {
-			parts.push(format!("SameSite={}", same_site));
-		}
-
-		parts.push(format!("Max-Age={}", self.config.ttl.as_secs()));
-
-		parts.join("; ")
-	}
-}
-
-impl Default for SessionMiddleware {
-	fn default() -> Self {
-		Self::with_defaults()
-	}
-}
-
-#[async_trait]
-impl Middleware for SessionMiddleware {
-	async fn process(&self, request: Request, handler: Arc<dyn Handler>) -> Result<Response> {
-		// Get or generate session ID
-		let session_id = self.get_session_id(&request);
-		let mut session = if let Some(id) = session_id.clone() {
-			self.store
-				.get(&id)
-				.filter(|s| s.is_valid())
-				.unwrap_or_else(|| SessionData::new(self.config.ttl))
-		} else {
-			SessionData::new(self.config.ttl)
-		};
-
-		// Touch the session
-		session.touch(self.config.ttl);
-
-		// Save the session
-		self.store.save(session.clone());
-
-		// Inject session ID and cookie name into request extensions
-		// so downstream handlers and Injectable impls can access them
-		request
-			.extensions
-			.insert(SessionId::new(session.id.clone()));
-		request
-			.extensions
-			.insert(SessionCookieName::new(self.config.cookie_name.clone()));
-		// Shared, mutable holder so handlers that rotate the session ID
-		// (`SessionData::regenerate_id`) keep `Set-Cookie` in sync. See #3827.
-		let active_id = ActiveSessionId::new(session.id.clone());
-		request.extensions.insert(active_id.clone());
-
-		// Call the handler
-		// Convert errors to responses so post-processing (e.g., security headers)
-		// always runs, even when invoked outside MiddlewareChain. (#3244)
-		let mut response = match handler.handle(request).await {
-			Ok(resp) => resp,
-			Err(e) => Response::from(e),
-		};
-
-		// Append Set-Cookie header (use append to preserve existing Set-Cookie headers).
-		// Read the final session ID from the shared holder rather than the
-		// local `session` clone, since handlers may have rotated the ID via
-		// `SessionData::regenerate_id`. See #3827.
-		let final_id = active_id.get();
-		let cookie = self.build_cookie_header(&final_id);
-		response.headers.append(
-			hyper::header::SET_COOKIE,
-			hyper::header::HeaderValue::from_str(&cookie).map_err(|e| {
-				reinhardt_core::exception::Error::Internal(format!(
-					"Failed to create cookie header: {}",
-					e
-				))
-			})?,
-		);
-
-		Ok(response)
-	}
-}
+//!
+//! This module is split into responsibility-focused submodules:
+//!
+//! - [`id`] — session-ID newtypes and the request-scoped active ID holder
+//! - [`data`] — the `SessionData` payload, read/write/rotate helpers
+//! - [`store`] — in-memory `SessionStore` with lazy eviction
+//! - [`backend`] — pluggable `AsyncSessionBackend` trait
+//! - [`config`] — `SessionConfig` cookie/TTL knobs
+//! - [`middleware`] — `SessionMiddleware` that wires it all together
+//! - [`injectable`] — DI integration (`Injectable` impls + `SessionStoreRef`)
+//!
+//! All public types are re-exported here so existing call sites that
+//! used `crate::session::*` continue to work unchanged.
+
+mod backend;
+mod config;
+mod data;
+mod id;
+mod injectable;
+mod middleware;
+mod store;
+
+pub use backend::AsyncSessionBackend;
+pub use config::SessionConfig;
+pub use data::SessionData;
+pub use id::{ActiveSessionId, SessionCookieName, SessionId};
+pub use injectable::SessionStoreRef;
+pub use middleware::SessionMiddleware;
+pub use store::SessionStore;
 
 #[cfg(test)]
 mod tests {
 	use super::*;
+	use async_trait::async_trait;
 	use bytes::Bytes;
 	use hyper::{HeaderMap, Method, StatusCode, Version};
+	use reinhardt_http::{Handler, Middleware, Request, Response, Result};
+	#[allow(deprecated)]
+	use reinhardt_conf::Settings;
+	use std::sync::{Arc, RwLock};
 	use std::thread;
+	use std::time::{Duration, SystemTime};
 
 	struct TestHandler;
 
@@ -1486,138 +801,14 @@ mod tests {
 	}
 }
 
-// ============================================================================
-// Injectable Implementations for Dependency Injection
-// ============================================================================
-
-/// Default session cookie name used when no `SessionCookieName` extension is present.
-const DEFAULT_SESSION_COOKIE_NAME: &str = "sessionid";
-
-/// Helper function to extract session ID from HTTP request cookies.
-///
-/// Searches for a cookie with the specified name in the Cookie header.
-///
-/// # Arguments
-///
-/// * `request` - The HTTP request to extract the session ID from
-/// * `cookie_name` - The name of the session cookie (e.g., "sessionid")
-///
-/// # Returns
-///
-/// * `Ok(String)` - The session ID if found and valid
-/// * `Err(DiError)` - If the cookie header is missing, invalid, or the session cookie is not found
-fn extract_session_id_from_request(request: &Request, cookie_name: &str) -> DiResult<String> {
-	let cookie_header = request
-		.headers
-		.get(hyper::header::COOKIE)
-		.ok_or_else(|| DiError::NotFound("Cookie header not found".to_string()))?;
-
-	let cookie_str = cookie_header
-		.to_str()
-		.map_err(|e| DiError::ProviderError(format!("Invalid cookie header: {}", e)))?;
-
-	for cookie in cookie_str.split(';') {
-		let parts: Vec<&str> = cookie.trim().splitn(2, '=').collect();
-		if parts.len() == 2 && parts[0] == cookie_name {
-			return Ok(parts[1].to_string());
-		}
-	}
-
-	Err(DiError::NotFound(format!(
-		"Session cookie '{}' not found",
-		cookie_name
-	)))
-}
-
-#[async_trait]
-impl Injectable for SessionData {
-	async fn inject(ctx: &InjectionContext) -> DiResult<Self> {
-		// Get SessionStore from SingletonScope
-		let store = ctx.get_singleton::<Arc<SessionStore>>().ok_or_else(|| {
-			DiError::NotFound(
-				"SessionStore not found in SingletonScope. \
-                     Ensure SessionMiddleware is configured and its store is registered."
-					.to_string(),
-			)
-		})?;
-
-		// Get Request from context
-		let request = ctx.get_request::<Request>().ok_or_else(|| {
-			DiError::NotFound("Request not found in InjectionContext".to_string())
-		})?;
-
-		// Extract configured cookie name from request extensions.
-		// Extensions::get returns an owned value, so we extract it once and
-		// use a reference for the lookup to avoid additional allocation.
-		let ext_cookie_name = request.extensions.get::<SessionCookieName>();
-		let cookie_name = ext_cookie_name
-			.as_ref()
-			.map(|cn| cn.as_str())
-			.unwrap_or(DEFAULT_SESSION_COOKIE_NAME);
-
-		// Prefer the SessionId injected by SessionMiddleware (present for all requests,
-		// including those without a Cookie header such as the initial login request).
-		// Fall back to parsing the Cookie header for requests that bypass the middleware.
-		let session_id = if let Some(sid) = request.extensions.get::<SessionId>() {
-			sid.as_ref().to_string()
-		} else {
-			extract_session_id_from_request(&request, cookie_name)?
-		};
-
-		// Load SessionData from store, attaching the request-scoped active session
-		// ID holder so `SessionData::regenerate_id` can keep the middleware's
-		// `Set-Cookie` value in sync with rotations. See #3827.
-		let id_holder = request.extensions.get::<ActiveSessionId>();
-		let mut session = store
-			.get(&session_id)
-			.filter(|s| s.is_valid())
-			.ok_or_else(|| {
-				DiError::NotFound("Valid session not found. Session may have expired.".to_string())
-			})?;
-		session.id_holder = id_holder;
-		Ok(session)
-	}
-}
-
-/// Wrapper for `Arc<SessionStore>` to enable dependency injection
-///
-/// This wrapper type is necessary because we cannot implement Injectable
-/// for `Arc<SessionStore>` directly due to Rust's orphan rules.
-#[derive(Clone)]
-pub struct SessionStoreRef(pub Arc<SessionStore>);
-
-impl SessionStoreRef {
-	/// Get a reference to the inner SessionStore
-	pub fn inner(&self) -> &SessionStore {
-		&self.0
-	}
-
-	/// Get a clone of the inner `Arc<SessionStore>`
-	pub fn arc(&self) -> Arc<SessionStore> {
-		Arc::clone(&self.0)
-	}
-}
-
-#[async_trait]
-impl Injectable for SessionStoreRef {
-	async fn inject(ctx: &InjectionContext) -> DiResult<Self> {
-		ctx.get_singleton::<Arc<SessionStore>>()
-			.map(|arc_store| SessionStoreRef(Arc::clone(&*arc_store)))
-			.ok_or_else(|| {
-				DiError::NotFound(
-					"SessionStore not found in SingletonScope. \
-                     Ensure SessionMiddleware is configured and its store is registered."
-						.to_string(),
-				)
-			})
-	}
-}
-
 #[cfg(test)]
 mod async_backend_tests {
 	use super::*;
+	use async_trait::async_trait;
+	use reinhardt_http::Result;
 	use std::collections::HashMap;
 	use std::sync::{Arc, RwLock};
+	use std::time::Duration;
 
 	/// In-memory MockBackend for testing `AsyncSessionBackend`.
 	struct MockBackend {
